[feat/#14] 나머지 엔티티 레포지토리 구현
