+++ src/main/java/com/clokey/server/domain/ClothImage/dao/ClothImageRepository.java
@@ -1,0 +1,7 @@
+package com.clokey.server.domain.ClothImage.dao;
+
+import com.clokey.server.domain.model.ClothImage;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ClothImageRepository extends JpaRepository<ClothImage, Long> {
+}
