+++ src/main/java/com/clokey/server/domain/HashtagHistory/dao/HashtagHistoryRepository.java
@@ -1,0 +1,7 @@
+package com.clokey.server.domain.HashtagHistory.dao;
+
+import com.clokey.server.domain.model.mapping.HashtagHistory;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface HashtagHistoryRepository extends JpaRepository<HashtagHistory, Long> {
+}
