+++ src/main/java/com/clokey/server/domain/HistoryImage/dao/HistoryImageRepository.java
@@ -1,0 +1,7 @@
+package com.clokey.server.domain.HistoryImage.dao;
+
+import com.clokey.server.domain.model.HistoryImage;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface HistoryImageRepository extends JpaRepository<HistoryImage, Long> {
+}
