--- src/main/java/com/clokey/server/domain/model/Follow.java
@@ -1,27 +1,0 @@
-package com.clokey.server.domain.model;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Table(uniqueConstraints = @UniqueConstraint(columnNames = {"following_user_id", "followed_user_id"}))
-public class Follow {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "following_user_id", nullable = false)
-    private Member following;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "followed_user_id", nullable = false)
-    private Member followed;
-
-}
-
