+++ src/main/java/com/clokey/server/domain/model/Term.java
@@ -1,11 +1,7 @@
 package com.clokey.server.domain.model;
 
-import com.clokey.server.domain.model.mapping.MemberAgree;
 import jakarta.persistence.*;
 import lombok.*;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Getter
